GH-34603: [Go][Parquet] Problem writing dictionary with empty strings (#34709)
Writing a dictionary encoded column consisting of empty strings ended up writing values that consisted of a string with a NUL character in them rather than actually writing an empty string. This fixes that issue and also cleans the code up a little bit in doing so.

A unit test is added to test for the behavior.

Users who wrote dictionary ByteArray or FixedLenByteArray columns that contained empty strings will see this fixed when it comes to handling those empty strings rather than having written strings containing a single NUL character (`\x00`).

* Closes: #34603

Authored-by: Matt Topol <zotthewizard@gmail.com>
Signed-off-by: Matt Topol <zotthewizard@gmail.com>